일단 ret을 double형으로 바꿔봄

+++ src/main/java/calculator/ACalculator/AddOperator.java
@@ -2,7 +2,7 @@ package calculator.ACalculator;
 
 public class AddOperator  implements InterOperator {
     @Override
-    public int operate(int a, int b) {
+    public double operate(int a, int b) {
         return a+b;
     }
 }
