+++ src/main/java/calculator/ACalculator/InterOperator.java
@@ -1,5 +1,5 @@
 package calculator.ACalculator;
 
 public interface InterOperator {
-    int operate(int a, int b);
+    double operate(int a, int b);
 }
