+++ src/main/java/calculator/ACalculator/ModOperator.java
@@ -3,10 +3,10 @@ package calculator.ACalculator;
 public class ModOperator implements InterOperator{
 
     @Override
-    public int operate(int a, int b) {
-        int ret = 0;
+    public double operate(int a, int b) {
+        double ret = 0;
         try{
-            ret = a%b;
+            ret = a%(double)b;
         }
         catch(ArithmeticException e){
             System.out.println("0은 분모에 위치할 수 없습니다.");
