+++ src/main/java/calculator/ACalculator/MultiplyOperator.java
@@ -2,7 +2,7 @@ package calculator.ACalculator;
 
 public class MultiplyOperator implements InterOperator {
     @Override
-    public int operate(int a, int b) {
+    public double operate(int a, int b) {
         return a*b;
     }
 }
