+++ src/main/java/calculator/ACalculator/SubtractOperator.java
@@ -2,7 +2,7 @@ package calculator.ACalculator;
 
 public class SubtractOperator  implements InterOperator {
     @Override
-    public int operate(int a, int b) {
+    public double operate(int a, int b) {
         return a-b;
     }
 }
